NEXUS-4149: fix regarding "descriptor" document and added low-level test to validate transport file correctness and content.
863469cf34f2da29a5aa24f3e158fb7abf6ef6c3

NEXUS-4149: fixed.

The culprit was the MergedIndexingContext, that when published, resulted in "transport format" (the GZ file) that contained multiple allGroups and rootGroups fields, the one used in "tree view" components (to list root nodes of the index tree). Nexus handled (1.8 tested) and handles this nicely, but m2e was not. It was simply finding _one_ of the present allGroups field (usually the wrong), and presenting tree using it, while the Lucene index itself was correct, and was usable for searches. This affects only "tree view", not the searches.

bb8f61fde52a78d85aaab228ab6a239aa9faa1dd